# pkg/apidoc/openapi.yaml
...
openapi: 3.0.3
info:
  title: expense-service API
  version: "1.0.0"
  description: |
    JSON API of **expense-service**, a small expense-logging service (Go + Gin + SQLite).
    It has no authentication yet. Run it on 127.0.0.1 only.

    ## Conventions

    - **Identifiers:** every resource is addressed **only by `uid`**, a server-assigned,
      immutable UUID v4 (lowercase in responses; path uids are matched case-insensitively).
      Internal numeric row ids never appear in URLs or JSON. A path `{uid}` that is not
      UUID-shaped (e.g. `/api/transactions/1`) is **400**; an unknown uid is **404**.
    - **Amounts** are exact. Clients send a decimal **string** (`"120.50"`, recommended) or a
      JSON number, which is read as its literal text and never as a float. Responses carry both the
      decimal string (`amount`, `balance`, `debt`, …) and the **integer minor units**
      (`amount_minor`, `balance_minor`, …). The number of decimals depends on the currency (ISO 4217):
      THB/USD 2, JPY/KRW 0, KWD/BHD 3, CLF 4. Input with more decimals than the currency
      allows is rejected (422), not rounded. Extra trailing zeros are fine (`"1500.00"` JPY).
      There is **no currency conversion** anywhere.
    - **Time:** `spent_at` is an RFC 3339 / ISO 8601 date-time **with an offset**
      (`2026-10-06T21:06:00+08:00` or `…Z`). Any offset is accepted. The value is stored at second
      precision with the offset as entered. `created_at` / `updated_at` are UTC (`…Z`).
...

...
servers:
...
paths:
...
  /api/balances:
    post:
      tags: [balances]
      operationId: createBalance
      summary: Create a balance
      description: |
        Assets (`payment_account`, `other_asset`) need `balance` (may be negative). Liabilities
        (`credit_card`, `other_liability`) need `debt` (may be negative, i.e. in credit) and `limit` (≥ 0).
        Sending an amount that does not apply to the type is a 422. Names are unique case-insensitively (409).
        A `version` in the body is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BalanceInput' }
            examples:
              payment_account:
                summary: Bank account
                value: { name: KBank debit, type: payment_account, currency: THB, balance: '10000' }
              credit_card:
                summary: Credit card
                value: { name: KBank Visa, type: credit_card, currency: THB, debt: '52000', limit: '50000' }
      responses:
        '201':
          description: Created.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/LocationBalance' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Balance' }
              examples:
                payment_account: { $ref: '#/components/examples/BalanceAsset' }
                credit_card: { $ref: '#/components/examples/BalanceCard' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: A balance with this name already exists (case-insensitive).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                duplicate_name: { $ref: '#/components/examples/ErrBalanceDuplicateName' }
        '422': { $ref: '#/components/responses/BalanceValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [balances]
      operationId: listBalances
      summary: List balances
      description: |
        Grouped by type (payment accounts, credit cards, other assets, other liabilities), then by
        name. `totals` sums the returned rows per currency.
      parameters:
        - name: type
          in: query
          description: Only balances of this type (case-insensitive). Cannot be combined with `payable`.
          schema: { $ref: '#/components/schemas/BalanceType' }
        - name: payable
          in: query
          description: |
            `1` or `true` (case-insensitive) returns only `payment_account` and `credit_card`
            balances. **Any other value is treated as "not set"** (not a 400).
          schema: { type: string, example: '1' }
      responses:
        '200':
          description: The balances and per-currency totals.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BalanceList' }
        '400':
          description: Invalid query parameters (`type` unknown, or `type` combined with `payable`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                type:
                  value:
                    error: invalid query parameters
                    fields: { type: 'must be one of payment_account, credit_card, other_asset, other_liability' }
                combined:
                  value:
                    error: invalid query parameters
                    fields: { payable: cannot be combined with type }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/balances/{uid}:
    parameters:
      - $ref: '#/components/parameters/BalanceUid'
    get:
      tags: [balances]
      operationId: getBalance
      summary: Get a balance
      responses:
        '200':
          description: The balance. `ETag` carries its version.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Balance' }
              examples:
                payment_account: { $ref: '#/components/examples/BalanceAsset' }
                credit_card: { $ref: '#/components/examples/BalanceCard' }
        '400': { $ref: '#/components/responses/BadUid' }
        '404': { $ref: '#/components/responses/BalanceNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [balances]
      operationId: replaceBalance
      summary: Replace a balance (manual override)
      description: |
        Full replace with the create rules, except that `type` is immutable: it is still required
        and must equal the stored type (422 on `type` otherwise). The amounts sent override whatever
        transactions did to the balance. A change of `balance` (assets) or `debt` (liabilities) records
        a read-only `balance_adjustment` transaction. Limit-only, name, description and currency changes
        record nothing. The currency cannot change while transactions reference the balance (409
        `balance_in_use`). An omitted `description` is cleared. Requires `If-Match` or `"version"`.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BalanceInput' }
            example: { name: KBank debit, type: payment_account, currency: THB, description: Salary account, balance: '5000', version: 3 }
      responses:
        '200':
          description: Updated. `ETag` carries the new version.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Balance' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/BalanceNotFound' }
        '409': { $ref: '#/components/responses/BalanceConflict' }
        '422': { $ref: '#/components/responses/BalanceValidationFailed' }
        '428': { $ref: '#/components/responses/VersionRequired' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [balances]
      operationId: deleteBalance
      summary: Delete a balance
      description: |
        Deletes the balance. `If-Match` is optional; a stale one is a 409. Transactions that
        reference the balance are **not** blocked or deleted: they keep its uid and name snapshot, and
        later edits or deletes of them skip the missing balance.
      parameters:
        - $ref: '#/components/parameters/IfMatchOptional'
      responses:
        '204': { description: Deleted. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/BalanceNotFound' }
        '409':
          description: Stale `If-Match`. The body has the current balance and the response has its `ETag`.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BalanceConflictError' }
              examples:
                version_conflict: { $ref: '#/components/examples/ErrBalanceVersionConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
...
    BalanceUid:
      name: uid
      in: path
      required: true
      description: Balance uid (UUID, case-insensitive). Numeric ids are not accepted (400).
      schema: { $ref: '#/components/schemas/Uid' }
      example: 4dc73a4a-a60b-414e-98bc-3c71298a0af5
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: |
        The version this write is based on, as an entity tag: the `ETag` (or `version`) of your last
        read, e.g. `"3"` (`W/"3"` also accepted; `*` and lists are not). Either this header or
        `"version"` in the body is **required** (428 if neither is sent). If both are sent they must agree (400).
        A stale version is a 409.
      schema: { type: string, example: '"3"' }
    IfMatchOptional:
      name: If-Match
      in: header
      required: false
      description: Optional. The `ETag` of your last read, e.g. `"3"`. If sent and stale the delete is refused with 409; if malformed, 400.
      schema: { type: string, example: '"3"' }
  headers:
    ETag:
      description: 'The record''s `version` as a strong entity tag, e.g. `"3"`. Send it back in `If-Match`.'
      schema: { type: string, example: '"1"' }
...
    LocationBalance:
      description: URL of the new balance.
      schema: { type: string, example: /api/balances/4dc73a4a-a60b-414e-98bc-3c71298a0af5 }
...
  responses:
    BadRequest:
      description: |
        Malformed request: invalid JSON, a body that is not one JSON object, unknown field, wrong JSON
        type, empty users patch, body over 64 KiB, a uid that is not UUID-shaped, or a malformed / disagreeing `If-Match`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            malformed_json: { value: { error: malformed JSON at byte 12 } }
            unknown_field: { value: { error: 'unknown field "account"' } }
            wrong_type: { value: { error: 'field "note" must be a string' } }
            bad_uid: { $ref: '#/components/examples/ErrBadUid' }
            bad_if_match: { value: { error: 'If-Match must be a single version entity tag such as "3" (the ETag header or "version" field of your last read)' } }
            versions_disagree: { value: { error: 'If-Match ("3") and the body "version" (2) disagree; send one, or the same value in both' } }
    BadUid:
      description: The path uid is not UUID-shaped (e.g. a numeric id).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            bad_uid: { $ref: '#/components/examples/ErrBadUid' }
...
    BalanceNotFound:
      description: No balance has this uid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: balance not found }
...
    BalanceValidationFailed:
      description: Validation failed (per-type amount rules, immutable `type`, unknown currency, …).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
          examples:
            type_changed: { value: { error: validation failed, fields: { type: balance type cannot be changed after creation } } }
            amount_rules:
              value:
                error: validation failed
                fields:
                  debt: is required for type credit_card
                  balance: is not used for type credit_card (only for payment_account and other_asset)
            decimals: { value: { error: validation failed, fields: { balance: has too many decimal places (max 2 for this currency) } } }
    VersionRequired:
      description: Neither `If-Match` nor `"version"` was sent.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: 'this update needs the version it is based on: send an If-Match header with the ETag of your last read (e.g. If-Match: "3") or "version" in the body'
            code: version_required
...
    BalanceConflict:
      description: |
        `version_conflict` (stale version; `current` + `ETag` sent), `balance_in_use` (currency change
        while transactions reference the balance), or a duplicate name (no `code`).
      headers:
        ETag:
          description: Current version (only with `version_conflict`).
          schema: { type: string, example: '"3"' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BalanceConflictError' }
          examples:
            version_conflict: { $ref: '#/components/examples/ErrBalanceVersionConflict' }
            balance_in_use: { $ref: '#/components/examples/ErrBalanceInUse' }
            duplicate_name: { $ref: '#/components/examples/ErrBalanceDuplicateName' }
...
    InternalError:
      description: Unexpected server error (details are only logged).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: internal server error }
  schemas:
    Uid:
      type: string
      format: uuid
      description: Server-assigned UUID v4. Lowercase in responses; accepted in any case.
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      example: 4dc73a4a-a60b-414e-98bc-3c71298a0af5
    Currency:
      type: string
      description: ISO 4217 code (trimmed and upper-cased on input; responses are upper-case). Determines the minor-unit scale.
      pattern: '^[A-Za-z]{3}$'
      example: THB
    DecimalInput:
      description: |
        Plain decimal as a JSON string (recommended) or number, read as text (never a float):
        no signs (except where negative values are allowed), commas, symbols or exponents.
        At most 13 digits before the point, and at most as many decimals as the currency allows.
      oneOf:
        - type: string
          example: '120.50'
        - type: number
          example: 120.5
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 in UTC with second precision.
      example: '2026-10-08T13:28:56Z'
    Version:
      type: integer
      format: int64
      minimum: 1
      description: Optimistic-locking version (1 on create, +1 on every change).
      example: 1
    Error:
      type: object
      description: Shape of every error response.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message.
        fields:
          type: object
          description: Field name → problem (validation, query, conflict and duplicate errors).
          additionalProperties: { type: string }
        code:
          type: string
          description: Stable machine-readable reason, present for some errors only.
          enum: [version_required, version_conflict, balance_in_use, balance_adjustment_readonly]
        current:
          type: object
          description: With `version_conflict` only, the record (Transaction or Balance) as it is now.
    ValidationError:
      type: object
      required: [error, fields]
      properties:
        error: { type: string, enum: [validation failed] }
        fields:
          type: object
          additionalProperties: { type: string }
...
    BalanceConflictError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [version_conflict, balance_in_use]
          description: Absent for a duplicate name.
        fields:
          type: object
          additionalProperties: { type: string }
        current: { $ref: '#/components/schemas/Balance' }
...
    BalanceType:
      type: string
      description: '`payment_account` and `other_asset` are assets (they have `balance`); `credit_card` and `other_liability` are liabilities (they have `debt` and `limit`).'
      enum: [payment_account, credit_card, other_asset, other_liability]
    Balance:
      type: object
      description: |
        Fields that do not apply to the type are `null`: assets have `balance`; liabilities have
        `debt`, `limit`, `available` and `over_limit`.
      required: [uid, name, type, kind, currency, description, balance, balance_minor, debt, debt_minor, limit, limit_minor, available, available_minor, over_limit, created_at, updated_at, version]
      properties:
        uid: { $ref: '#/components/schemas/Uid' }
        name: { type: string, maxLength: 100, example: KBank debit }
        type: { $ref: '#/components/schemas/BalanceType' }
        kind:
          type: string
          enum: [asset, liability]
          description: Derived from `type`.
        currency: { type: string, pattern: '^[A-Z]{3}$', example: THB }
        description: { type: string, maxLength: 1000, example: '' }
        balance:
          type: string
          nullable: true
          description: Assets only. May be negative (overdraft).
          example: '10000.00'
        balance_minor: { type: integer, format: int64, nullable: true, example: 1000000 }
        debt:
          type: string
          nullable: true
          description: Liabilities only. May be negative (in credit).
          example: null
        debt_minor: { type: integer, format: int64, nullable: true, example: null }
        limit:
          type: string
          nullable: true
          description: Liabilities only; ≥ 0.
          example: null
        limit_minor: { type: integer, format: int64, nullable: true, example: null }
        available:
          type: string
          nullable: true
          description: limit − debt (liabilities only). Negative when over the limit.
          example: null
        available_minor: { type: integer, format: int64, nullable: true, example: null }
        over_limit:
          type: boolean
          nullable: true
          description: debt > limit (liabilities only).
          example: null
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: UTC; `null` until changed manually or by a transaction.
          example: null
        version: { $ref: '#/components/schemas/Version' }
    BalanceInput:
      type: object
      description: |
        Body of POST and PUT. Amounts that are absent, `null` or `""` count as not provided. Unknown
        fields are a 400. This includes read-only response fields such as `kind`, `*_minor`, `available`,
        `over_limit`, `created_at` and `updated_at`.
      required: [name, type, currency]
      properties:
        name:
          type: string
          maxLength: 100
          description: Trimmed; unique case-insensitively.
          example: KBank debit
        type:
          allOf:
            - $ref: '#/components/schemas/BalanceType'
          description: Required. On PUT it must equal the stored type (immutable).
        currency: { $ref: '#/components/schemas/Currency' }
        description: { type: string, nullable: true, maxLength: 1000 }
        balance:
          allOf:
            - $ref: '#/components/schemas/DecimalInput'
          nullable: true
          description: Required for assets (may be negative). Not allowed for liabilities.
        debt:
          allOf:
            - $ref: '#/components/schemas/DecimalInput'
          nullable: true
          description: Required for liabilities (may be negative). Not allowed for assets.
        limit:
          allOf:
            - $ref: '#/components/schemas/DecimalInput'
          nullable: true
          description: Required for liabilities (≥ 0). Not allowed for assets.
        uid:
          type: string
          description: Accepted and ignored.
        version:
          type: integer
          format: int64
          minimum: 1
          description: PUT only, as an alternative to `If-Match`. Ignored on POST.
      additionalProperties: false
    BalanceTotals:
      type: object
      description: Sums for one currency over the returned balances (decimal strings only; no conversion).
      required: [currency, assets, liabilities, net, credit_limit, available_credit]
      properties:
        currency: { type: string, example: THB }
        assets: { type: string, description: Sum of asset balances., example: '10000.00' }
        liabilities: { type: string, description: Sum of liability debts., example: '52000.00' }
        net: { type: string, description: assets − liabilities., example: '-42000.00' }
        credit_limit: { type: string, description: Sum of liability limits., example: '50000.00' }
        available_credit: { type: string, description: credit_limit − liabilities., example: '-2000.00' }
    BalanceList:
      type: object
      required: [balances, count, totals]
      properties:
        balances:
          type: array
          items: { $ref: '#/components/schemas/Balance' }
        count: { type: integer, example: 2 }
        totals:
          type: array
          description: One entry per currency, in alphabetical order.
          items: { $ref: '#/components/schemas/BalanceTotals' }
...
  examples:
...
    BalanceAsset:
      summary: Payment account (asset)
      value:
        uid: 4dc73a4a-a60b-414e-98bc-3c71298a0af5
        name: KBank debit
        type: payment_account
        kind: asset
        currency: THB
        description: ''
        balance: '10000.00'
        balance_minor: 1000000
        debt: null
        debt_minor: null
        limit: null
        limit_minor: null
        available: null
        available_minor: null
        over_limit: null
        created_at: '2026-10-08T13:28:56Z'
        updated_at: null
        version: 1
    BalanceCard:
      summary: Credit card over its limit (liability)
      value:
        uid: a8465748-a48b-426a-9466-9a6706d8143a
        name: KBank Visa
        type: credit_card
        kind: liability
        currency: THB
        description: ''
        balance: null
        balance_minor: null
        debt: '52000.00'
        debt_minor: 5200000
        limit: '50000.00'
        limit_minor: 5000000
        available: '-2000.00'
        available_minor: -200000
        over_limit: true
        created_at: '2026-10-08T13:28:56Z'
        updated_at: null
        version: 1
...
    ErrBadUid:
      summary: Not a UUID (e.g. a numeric id)
      value: { error: uid must be a UUID (xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx) }
...
    ErrBalanceVersionConflict:
      summary: Stale version
      value:
        error: 'version conflict: this balance was changed since you read it (now at version 3); nothing was saved. Reload it, re-apply your change and send the new version'
        fields: { version: is stale; the current version is 3 }
        code: version_conflict
        current:
          uid: 4dc73a4a-a60b-414e-98bc-3c71298a0af5
          name: KBank debit
          type: payment_account
          kind: asset
          currency: THB
          description: ''
          balance: '4879.50'
          balance_minor: 487950
          debt: null
          debt_minor: null
          limit: null
          limit_minor: null
          available: null
          available_minor: null
          over_limit: null
          created_at: '2026-10-08T13:28:56Z'
          updated_at: '2026-10-08T13:29:10Z'
          version: 3
    ErrBalanceInUse:
      summary: Currency change while transactions reference the balance
      value:
        error: balance is used by transactions (2 transaction(s)); its currency cannot change
        fields: { currency: cannot change while transactions reference this balance (no exchange-rate conversion) }
        code: balance_in_use
    ErrBalanceDuplicateName:
      summary: Name already used
      value:
        error: a balance with this name already exists
        fields: { name: is already used by another balance (names are case-insensitive) }
...